>>> arreglado

+++ src/main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ojimenez <ojimenez@student.42barcelona.    +#+  +:+       +#+        */
+/*   By: agrimald <agrimald@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/09 17:46:28 by agrimald          #+#    #+#             */
-/*   Updated: 2024/01/29 21:46:12 by ojimenez         ###   ########.fr       */
+/*   Updated: 2024/02/02 17:33:54 by agrimald         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,30 +38,40 @@ int	main(int argc, char *argv[], char *env[])
 		{
 			//ft_exit(input);
 			err = parser(&tokens, input, e->env_cpy);
+			//print_tokens(tokens);
 			e->env_cpy = env;
 			exp = expander(tokens, e, &exec);
-
-			/*t_expander *nodo;
-			nodo = exp;
-			int i = 1;
-			int j = 0;
-			while (nodo)
+			if (exp == NULL)
 			{
-				printf("Nodo %d:\nTipo:%d\nLen = %d\n", i, nodo->exp_type, nodo->len);
-				while (j < nodo->len + 1)
+				printf("Error aqui\n");
+				continue;
+			}
+			printf("ERROR = %d\n", err);
+			if (err == 0)
+			{
+				/*t_expander *nodo;
+				nodo = exp;
+				int i = 1;
+				int j = 0;
+					
+				while (nodo)
 				{
-					printf("%s\n", nodo->exp_matr[j]);
-					j++;
-				}
-				printf("Token = %s\n", nodo->token);
-				j = 0;
-				i++;
-				nodo = nodo->next;
-			}*/
-			if (exp->exp_matr[0])
-				err = executor(exp, e, tokens, &exec);
-			destroy_all(&tokens, &exp);
-		}
+					printf("Nodo %d:\nTipo:%d\nLen = %d\n", i, nodo->exp_type, nodo->len);
+					while (j < nodo->len + 1)
+					{
+						printf("%s\n", nodo->exp_matr[j]);
+						j++;
+					}
+					printf("Token = %s\n", nodo->token);
+					j = 0;
+					i++;
+					nodo = nodo->next;
+				}*/
+				if (exp->exp_matr[0])
+					err = executor(exp, e, tokens, &exec);
+				destroy_all(&tokens, &exp);
+			}
+		}	
 		else
 			free(input);
 	}
